Refactored Questions directory code
